create `AccessibleViewRegistry`, move all applicable views/accessibility help dialogs over to use it (#212138)
