+++ src/vs/platform/accessibility/browser/accessibleView.ts
@@ -1,0 +1,166 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { createDecorator } from 'vs/platform/instantiation/common/instantiation';
+import { IKeyboardEvent } from 'vs/platform/keybinding/common/keybinding';
+import { IPickerQuickAccessItem } from 'vs/platform/quickinput/browser/pickerQuickAccess';
+import { Event } from 'vs/base/common/event';
+import { IAction } from 'vs/base/common/actions';
+
+export const IAccessibleViewService = createDecorator<IAccessibleViewService>('accessibleViewService');
+
+export const enum AccessibleViewProviderId {
+	Terminal = 'terminal',
+	TerminalChat = 'terminal-chat',
+	TerminalHelp = 'terminal-help',
+	DiffEditor = 'diffEditor',
+	Chat = 'panelChat',
+	InlineChat = 'inlineChat',
+	InlineCompletions = 'inlineCompletions',
+	KeybindingsEditor = 'keybindingsEditor',
+	Notebook = 'notebook',
+	Editor = 'editor',
+	Hover = 'hover',
+	Notification = 'notification',
+	EmptyEditorHint = 'emptyEditorHint',
+	Comments = 'comments'
+}
+
+export const enum AccessibleViewType {
+	Help = 'help',
+	View = 'view'
+}
+
+export const enum NavigationType {
+	Previous = 'previous',
+	Next = 'next'
+}
+
+export interface IAccessibleViewOptions {
+	readMoreUrl?: string;
+	/**
+	 * Defaults to markdown
+	 */
+	language?: string;
+	type: AccessibleViewType;
+	/**
+	 * By default, places the cursor on the top line of the accessible view.
+	 * If set to 'initial-bottom', places the cursor on the bottom line of the accessible view and preserves it henceforth.
+	 * If set to 'bottom', places the cursor on the bottom line of the accessible view.
+	 */
+	position?: 'bottom' | 'initial-bottom';
+	/**
+	 * @returns a string that will be used as the content of the help dialog
+	 * instead of the one provided by default.
+	 */
+	customHelp?: () => string;
+	/**
+	 * If this provider might want to request to be shown again, provide an ID.
+	 */
+	id?: AccessibleViewProviderId;
+}
+
+
+export interface IAccessibleViewContentProvider extends IBasicContentProvider {
+	id: AccessibleViewProviderId;
+	verbositySettingKey: string;
+	/**
+	 * Note that a Codicon class should be provided for each action.
+	 * If not, a default will be used.
+	 */
+	onKeyDown?(e: IKeyboardEvent): void;
+	/**
+	 * When the language is markdown, this is provided by default.
+	 */
+	getSymbols?(): IAccessibleViewSymbol[];
+	/**
+	 * Note that this will only take effect if the provider has an ID.
+	 */
+	onDidRequestClearLastProvider?: Event<AccessibleViewProviderId>;
+}
+
+
+export interface IAccessibleViewSymbol extends IPickerQuickAccessItem {
+	markdownToParse?: string;
+	firstListItem?: string;
+	lineNumber?: number;
+	endLineNumber?: number;
+}
+
+export interface IPosition {
+	lineNumber: number;
+	column: number;
+}
+
+export interface IAccessibleViewService {
+	readonly _serviceBrand: undefined;
+	show(provider: AccesibleViewContentProvider, position?: IPosition): void;
+	showLastProvider(id: AccessibleViewProviderId): void;
+	showAccessibleViewHelp(): void;
+	next(): void;
+	previous(): void;
+	navigateToCodeBlock(type: 'next' | 'previous'): void;
+	goToSymbol(): void;
+	disableHint(): void;
+	getPosition(id: AccessibleViewProviderId): IPosition | undefined;
+	setPosition(position: IPosition, reveal?: boolean): void;
+	getLastPosition(): IPosition | undefined;
+	/**
+	 * If the setting is enabled, provides the open accessible view hint as a localized string.
+	 * @param verbositySettingKey The setting key for the verbosity of the feature
+	 */
+	getOpenAriaHint(verbositySettingKey: string): string | null;
+	getCodeBlockContext(): ICodeBlockActionContext | undefined;
+}
+
+
+export interface ICodeBlockActionContext {
+	code: string;
+	languageId?: string;
+	codeBlockIndex: number;
+	element: unknown;
+}
+
+export type AccesibleViewContentProvider = AdvancedContentProvider | ExtensionContentProvider;
+
+export class AdvancedContentProvider implements IAccessibleViewContentProvider {
+
+	constructor(
+		public id: AccessibleViewProviderId,
+		public options: IAccessibleViewOptions,
+		public provideContent: () => string,
+		public onClose: () => void,
+		public verbositySettingKey: string,
+		public actions?: IAction[],
+		public next?: () => void,
+		public previous?: () => void,
+		public onKeyDown?: (e: IKeyboardEvent) => void,
+		public getSymbols?: () => IAccessibleViewSymbol[],
+		public onDidRequestClearLastProvider?: Event<AccessibleViewProviderId>,
+	) { }
+}
+
+export class ExtensionContentProvider implements IBasicContentProvider {
+
+	constructor(
+		public readonly id: string,
+		public options: IAccessibleViewOptions,
+		public provideContent: () => string,
+		public onClose: () => void,
+		public next?: () => void,
+		public previous?: () => void,
+		public actions?: IAction[],
+	) { }
+}
+
+export interface IBasicContentProvider {
+	id: string;
+	options: IAccessibleViewOptions;
+	onClose(): void;
+	provideContent(): string;
+	actions?: IAction[];
+	previous?(): void;
+	next?(): void;
+}
