+++ src/vs/workbench/contrib/comments/browser/comments.contribution.ts
@@ -25,7 +25,11 @@ import { IEditorService } from 'vs/workbench/services/editor/common/editorServic
 import { IUriIdentityService } from 'vs/platform/uriIdentity/common/uriIdentity';
 import { revealCommentThread } from 'vs/workbench/contrib/comments/browser/commentsController';
 import { MarshalledCommentThreadInternal } from 'vs/workbench/common/comments';
-import { accessibleViewCurrentProviderId, accessibleViewIsShown, AccessibleViewProviderId } from 'vs/workbench/contrib/accessibility/browser/accessibilityConfiguration';
+import { accessibleViewCurrentProviderId, accessibleViewIsShown } from 'vs/workbench/contrib/accessibility/browser/accessibilityConfiguration';
+import { AccessibleViewProviderId } from 'vs/platform/accessibility/browser/accessibleView';
+import { AccessibleViewRegistry } from 'vs/platform/accessibility/browser/accessibleViewRegistry';
+import { CommentsAccessibleView } from 'vs/workbench/contrib/comments/browser/commentsAccessibleView';
+import { CommentsAccessibilityHelp } from 'vs/workbench/contrib/comments/browser/commentsAccessibility';
 
 registerAction2(class Collapse extends ViewAction<CommentsPanel> {
 	constructor() {
@@ -187,3 +191,6 @@ export class UnresolvedCommentsBadge extends Disposable implements IWorkbenchCon
 }
 
 Registry.as<IWorkbenchContributionsRegistry>(Extensions.Workbench).registerWorkbenchContribution(UnresolvedCommentsBadge, LifecyclePhase.Eventually);
+
+AccessibleViewRegistry.register(new CommentsAccessibleView());
+AccessibleViewRegistry.register(new CommentsAccessibilityHelp());
